ring: add the beginnings of a ring test.
Also, make some revisions to the ring implementation. Some corrections were necessary
since I hadn't been compiling up to this point, so the test motivated some obvious fixes.

I'm not pleased with the fact that I need to supply the generic type to each of the functional
options... it feels like supplying it to `New(...)` on ring creation should be sufficient.
I need to look into fixing this unpleasant boilerplate one way or another.