ring: beef up the unit tests.
Primarily this adds unit tests to validate functional correctness of the
queue implementation. It tests the SPSC, SPMC, MPSC, and MPMC behaviors.
It also tests the wraparound indexing, which can be a bit tricky.

I also added some benchmarks for testing performance, but seeing some odd
results on my WSL2 instance. Will have to revisit these to see if it's
something benchmark related or what.