buffer: fix grow semantics.
Previously, we were only correctly handling the case where the compact does the zero.
However, if we didn't need a compact, we were failing to zero the newly extended region.
This is now fixed.

Also, the implementation for `Grow` and `GrowAndZero` was commonized.

Lastly, the unit tests for the buffer implementation were fleshed out more.