ring: fix concurrency error in MPMC implementation.
The MPMC unit test ended up exposing an issue with the compare-and-swap semantics that we
were relying on. The DPDK implementation could get away with that due to its memory barriers,
but we do not have those, so it was creating race conditions between when the data was published
and when we marked it as available for reading.

Instead, we need to use atomic stores as the delineating point where a memory barrier would
normally go. This affects the implementation of `doDequeue`, `doEnqueue`, `moveProdHead`, and
`moveConsHead` in a significant way.

Now moving the head (prod or cons) marks the reservation. `doEnqueue` and `doDequeue` mark that
the data is published with an atomic store. As a plus side, this removes the CAS loops that we
had previously.

Also did some minor hygiene adjustments and added some relevant performance notes in light of the
changed concurrency semantics.